change HistogramSchema type.
