+++ commons/src/main/java/com/navercorp/pinpoint/common/trace/ServiceTypeCategory.java
@@ -24,7 +24,7 @@ public enum ServiceTypeCategory {
     SERVER((short)1000, (short)1999),
     DATABASE((short)2000, (short)2999),
     LIBRARY((short)5000, (short)7999),
-    CACHE_LIBRARY((short)8000, (short)8999, HistogramSchema.FAST_SCHEMA),
+    CACHE_LIBRARY((short)8000, (short)8999, BaseHistogramSchema.FAST_SCHEMA),
     RPC((short)9000, (short)9999);
    
     
@@ -33,7 +33,7 @@ public enum ServiceTypeCategory {
     private HistogramSchema histogramSchema;
 
     private ServiceTypeCategory(short minCode, short maxCode) {
-        this(minCode, maxCode, HistogramSchema.NORMAL_SCHEMA);
+        this(minCode, maxCode, BaseHistogramSchema.NORMAL_SCHEMA);
     }
     
     private ServiceTypeCategory(short minCode, short maxCode, HistogramSchema histogramSchema) {
