+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/receiver/service/ActiveThreadCountService.java
@@ -19,6 +19,7 @@
 
 package com.navercorp.pinpoint.profiler.receiver.service;
 
+import com.navercorp.pinpoint.common.trace.BaseHistogramSchema;
 import com.navercorp.pinpoint.common.trace.HistogramSchema;
 import com.navercorp.pinpoint.common.trace.HistogramSlot;
 import com.navercorp.pinpoint.common.trace.SlotType;
@@ -80,7 +81,7 @@ public class ActiveThreadCountService implements ProfilerRequestCommandService, 
 
     private final ActiveTraceLocator activeTraceLocator;
     private final int activeThreadSlotsCount;
-    private final HistogramSchema histogramSchema = HistogramSchema.NORMAL_SCHEMA;
+    private final HistogramSchema histogramSchema = BaseHistogramSchema.NORMAL_SCHEMA;
 
     public ActiveThreadCountService(ActiveTraceLocator activeTraceLocator) {
         this(activeTraceLocator, DEFAULT_FLUSH_DELAY);
